+++ st2common/st2common/models/api/tag.py
@@ -13,13 +13,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from st2common.persistence.base import Access
-from st2common.models.db.actionrunner import actionrunner_access
+from st2common.models.db.stormbase import TagField
 
 
-class ActionRunner(Access):
-    IMPL = actionrunner_access
+class TagsHelper(object):
 
-    @classmethod
-    def _get_impl(kls):
-        return kls.IMPL
+    @staticmethod
+    def to_model(tags):
+        return [TagField(name=tag.get('name', ''), value=tag.get('value', '')) for tag in tags]
+
+    @staticmethod
+    def from_model(tags):
+        return [{'name': tag.name, 'value': tag.value} for tag in tags]